chore(jans-cedarling): refactor, move `token_data` to authz module
