+++ jans-cedarling/cedarling/src/authz/entities/mod.rs
@@ -12,10 +12,9 @@ mod meta;
 #[cfg(test)]
 mod test_create;
 
-use crate::{
-    models::{cedar_schema::CedarSchemaJson, token_data::TokenPayload},
-    ResourceData,
-};
+use crate::authz::token_data::TokenPayload;
+use crate::{models::cedar_schema::CedarSchemaJson, ResourceData};
+
 use create::{create_entity, parse_namespace_and_typename, CedarPolicyCreateTypeError};
 
 /// Describe errors on creating entites for AccessToken
