+++ jans-cedarling/cedarling/src/authz/entities/test_create.rs
@@ -9,10 +9,8 @@
 
 use test_utils::assert_eq;
 
-use crate::models::{
-    cedar_schema::CedarSchemaJson,
-    token_data::{GetTokenClaimValue, TokenPayload},
-};
+use crate::authz::token_data::{GetTokenClaimValue, TokenPayload};
+use crate::models::cedar_schema::CedarSchemaJson;
 
 use super::create::*;
 use test_utils::SortedJson;
