+++ jans-cedarling/cedarling/src/models/mod.rs
@@ -12,4 +12,3 @@ pub(crate) mod app_types;
 pub(crate) mod cedar_schema;
 
 pub mod policy_store;
-pub mod token_data;
